Flatten client/panther sub-package into client
After GraphQL removal, the panther sub-package only held a constructor
and auth transport — no longer enough to justify a separate package.
Move NewRESTClient into rest.go (next to RESTClient), move authTransport
into transport.go, and delete internal/client/panther/ entirely.
