--- internal/client/panther/panther.go
@@ -1,36 +1,0 @@
-/*
-Copyright 2023 Panther Labs, Inc.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package panther
-
-import (
-	"strings"
-	"terraform-provider-panther/internal/client"
-)
-
-const graphQLPath = "/public/graphql"
-
-// NewRESTClient creates a configured REST client for the Panther API.
-func NewRESTClient(url, token string) *client.RESTClient {
-	// Strip the legacy /public/graphql suffix — older provider configs included it.
-	pantherURL := strings.TrimSuffix(url, graphQLPath)
-	httpClient := NewHTTPClient(token)
-
-	return &client.RESTClient{
-		Doer:    httpClient,
-		BaseURL: pantherURL,
-	}
-}
